ModelCache imported in this source tree.

+++ kdb/kodein-db/src/commonMain/kotlin/org/kodein/db/impl/model/cache/CachedModelSnapshot.kt
@@ -1,7 +1,8 @@
 package org.kodein.db.impl.model.cache
 
 import org.kodein.db.model.ModelDB
+import org.kodein.db.model.cache.ModelCache
 
-class CachedModelSnapshot(override val mdb: ModelDB.Snapshot, override val cache: ModelCache, override val cacheCopyMaxSize: Int) : BaseCachedModelRead, ModelDB.Snapshot {
+class CachedModelSnapshot(override val mdb: ModelDB.Snapshot, override val cache: ModelCache, override val cacheCopyMaxSize: Long) : BaseCachedModelRead, ModelDB.Snapshot {
     override fun close()  = mdb.close()
 }
